requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
anthropic==0.34.2
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
pyserial==3.5
